feat: Introduce helper module for token management and refactor authentication handlers to utilize it
